Started implementing Animation class, refactored package structure to remove Highlevel and Lowlevel packages. It still separates the two though
