+++ Silhouette/src/CSS/BaseComponents/Implementation/Statement.java
@@ -1,4 +1,4 @@
-package CSS.HighLevel.Implementation;
+package CSS.BaseComponents.Implementation;
 
 import java.util.ArrayList;
