--- Silhouette/src/CSS/HighLevel/Animation.java
@@ -1,5 +1,0 @@
-package CSS.HighLevel;
-
-public class Animation extends RuleSet {
-    //public void
-}
