+++ Silhouette/src/CSS/Values/Color.java
@@ -1,7 +1,4 @@
-package CSS.HighLevel.Values;
-
-import java.util.Arrays;
-import java.util.Iterator;
+package CSS.Values;
 
 public class Color {
     // Color
